backend/requirements.txt
aiofiles
aiohttp
asyncpraw
python-dotenv
imagehash
numpy
Pillow
tqdm